Fixed bug with sliders caused by migration to new message queue system

+++ Source/CabbageGUIClass.h
@@ -35,7 +35,7 @@
 
 class CabbageGUIClass : public CabbageUtils
 {
-        double width, height, top, left, isRect, min, max, tabpage, minX, minY, maxX, maxY, comboRange, fftSize, overlapSize, frameSize,
+        float width, height, top, left, isRect, min, max, tabpage, minX, minY, maxX, maxY, comboRange, fftSize, overlapSize, frameSize,
 		noOfMenus, onoff, midiChan, midiCtrl, sliderRange, xypadRangeY, xypadRangeX, noSteps, noPatterns, pvsChannel, alpha,
 		line, anchor, linkTo, scaleX, scaleY, value, valueX, valueY, maxItems, sliderIncr, sliderSkew, decimalPlaces, rCtrls, lineIsVertical;
         StringArray items, onoffcaptions, key, channels, snapshotData, colours;
